[REFACTOR/#111] entity 변경
