+++ app/src/main/java/com/napzak/market/domain/explore/model/ProductListFilter.kt
@@ -1,6 +1,7 @@
 package com.napzak.market.domain.explore.model
 
 data class ProductListFilter(
+    val storeId: Long? = null,
     val sortOption: String,
     val genreId: List<Long>?,
     val isOnSale: Boolean,
